requirements.txt
aiofiles==24.1.0
aiogram==3.21.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.10
MarkupSafe==3.0.2
middleware==1.2.3
multidict==6.6.3
openpyxl==3.1.5
propcache==0.3.2
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
python-telegram-bot==22.3
requests==2.32.4
sniffio==1.3.1
SQLAlchemy==2.0.41
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.1.3
yarl==1.20.1
